moved service implementation into its own module
